+++ src/core/assessment/AssessmentApiClient.ts
@@ -2,15 +2,12 @@ import appConfig from '@covid/appConfig';
 import { IApiClient } from '@covid/core/api/ApiClient';
 import { AssessmentInfosRequest } from '@covid/core/assessment//dto/AssessmentInfosRequest';
 import { AssessmentResponse } from '@covid/core/assessment//dto/AssessmentInfosResponse';
-import { VaccineRequest } from '@covid/core/vaccines/dto/VaccineRequest';
-import { VaccineResponse } from '@covid/core/vaccines/dto/VaccineResponse';
 
 const API_ASSESSMENTS = '/assessments/';
 
 export interface IAssessmentRemoteClient {
   addAssessment(assessment: AssessmentInfosRequest): Promise<AssessmentResponse>;
   updateAssessment(assessmentId: string, assessment: AssessmentInfosRequest): Promise<AssessmentResponse>;
-  saveVaccineResponse(patientId: string, payload: Partial<VaccineRequest>): Promise<VaccineResponse>;
 }
 
 export class AssessmentApiClient implements IAssessmentRemoteClient {
@@ -30,14 +27,5 @@ export class AssessmentApiClient implements IAssessmentRemoteClient {
   updateAssessment(assessmentId: string, assessment: AssessmentInfosRequest): Promise<AssessmentResponse> {
     const assessmentUrl = `/assessments/${assessmentId}/`;
     return this.apiClient.patch<AssessmentInfosRequest, AssessmentResponse>(assessmentUrl, assessment);
-  }
-
-  saveVaccineResponse(patientId: string, payload: Partial<VaccineRequest>): Promise<VaccineResponse> {
-    payload = {
-      ...payload,
-      patient: patientId,
-      version: appConfig.vaccineVersion,
-    };
-    return this.apiClient.post<VaccineRequest, VaccineResponse>('/vaccines/', payload as VaccineRequest);
   }
 }
