+++ src/core/assessment/AssessmentService.ts
@@ -1,5 +1,5 @@
 import { PatientInfosRequest } from '@covid/core/user/dto/UserAPIContracts';
-import { VaccineRequest } from '@covid/core/vaccines/dto/VaccineRequest';
+import { VaccineRequest } from '@covid/core/vaccine/dto/VaccineRequest';
 
 import { IAssessmentRemoteClient } from './AssessmentApiClient';
 import { IAssessmentState } from './AssessmentState';
@@ -81,10 +81,5 @@ export default class AssessmentService implements IAssessmentService {
 
     const response = this.sendFullAssessmentToApi();
     return !!response;
-  }
-
-  async saveVaccineResponse(patientId: string, payload: Partial<VaccineRequest>): Promise<boolean> {
-    await this.apiClient.saveVaccineResponse(patientId, payload);
-    return true;
   }
 }
