+++ src/core/vaccine/VaccineApiClient.ts
@@ -1,0 +1,25 @@
+import appConfig from '@covid/appConfig';
+import { IApiClient } from '@covid/core/api/ApiClient';
+import { VaccineRequest } from '@covid/core/vaccine/dto/VaccineRequest';
+import { VaccineResponse } from '@covid/core/vaccine/dto/VaccineResponse';
+
+
+export interface IVaccineRemoteClient {
+  saveVaccineResponse(patientId: string, payload: Partial<VaccineRequest>): Promise<VaccineResponse>;
+}
+
+export class VaccineApiClient implements IVaccineRemoteClient {
+  apiClient: IApiClient;
+  constructor(apiClient: IApiClient) {
+    this.apiClient = apiClient;
+  }
+
+  saveVaccineResponse(patientId: string, payload: Partial<VaccineRequest>): Promise<VaccineResponse> {
+    payload = {
+      ...payload,
+      patient: patientId,
+      version: appConfig.vaccineVersion,
+    };
+    return this.apiClient.post<VaccineRequest, VaccineResponse>('/vaccines/', payload as VaccineRequest);
+  }
+}
