+++ src/core/vaccine/dto/VaccineResponse.ts
