+++ src/features/assessment/index.ts
@@ -11,4 +11,3 @@ export { default as TreatmentSelectionScreen } from './TreatmentSelectionScreen'
 export { default as WhereAreYouScreen } from './WhereAreYouScreen';
 export { VaccineRegistryInfoScreen } from './gb/VaccineRegistryInfoScreen';
 export { VaccineRegistrySignUpScreen } from './gb/VaccineRegistrySignUpScreen';
-export { VaccineScreen } from './VaccineScreen';
