+++ src/features/index.ts
@@ -13,3 +13,4 @@ export * from './school-network';
 export { SplashScreen } from './SplashScreen';
 export * from './thank-you';
 export * from './validation-study';
+export * from './vaccines';
