+++ src/features/vaccines/index.ts
@@ -1,0 +1,1 @@
+export { VaccineScreen } from './VaccineScreen';
